fix: validate first char in input
This commit comes from the first fuzzing session I had.
An expected error but a good way to validate how fuzzing works.

This library did not validate the first char for the input. If it is not
a `{` we can terminated the tokenizer because it is against an invalid
input.